Mark component
